Adding Single responsability principle for blueprints and refactoring api configs
