+++ api.py
@@ -1,20 +1,12 @@
 # -*- coding: utf-8 -*-
-import os
 
-from flask import Flask, request
-from pymongo import MongoClient
-from bson import json_util
-from users.users import users_api, requer_autenticacao
+from flask import Flask
+from users.users import users_api
 from escolas.escolas import escolas_api
 from cardapios.cardapios import cardapios_api
 from editor.editor import editor_api
-from flasgger import Swagger, swag_from
+from flasgger import Swagger
 
-API_KEY = os.environ.get('API_KEY')
-API_MONGO_URI = 'mongodb://{}'.format(os.environ.get('API_MONGO_URI'))
-
-client = MongoClient(API_MONGO_URI)
-db = client['pratoaberto']
 
 def create_app():
 
@@ -26,6 +18,7 @@ def create_app():
     swagger = Swagger(app)
 
     return app
+
 
 if __name__ == '__main__':
     app = create_app()
