+++ settings/api_settings.py
@@ -1,0 +1,8 @@
+import os
+from pymongo import MongoClient
+
+API_KEY = os.environ.get('API_KEY')
+API_MONGO_URI = 'mongodb://{}'.format(os.environ.get('API_MONGO_URI'))
+
+client = MongoClient(API_MONGO_URI)
+db = client['pratoaberto']
